Fruit-freshness-insight/requirements.txt
tensorflow==2.17.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
pandas==2.1.3
plotly==5.18.0
numpy==1.26.2
flask==3.1.2
